clean up and remake/impliment
