--- demo_sessions/0putP7EhNeA2LZrvmxxo4XrUAOUtmDkkgcrS0y1Q2xY.json
@@ -1,1 +1,0 @@
-{"session_id": "0putP7EhNeA2LZrvmxxo4XrUAOUtmDkkgcrS0y1Q2xY", "user_id": "local-user", "issued_at": "2025-08-13T05:32:39.767574+00:00", "expires_at": "2025-08-13T13:32:39.767574+00:00", "last_seen": "2025-08-13T05:32:39.771726+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "nA6iosudMSFJrl4LF0dIjqioas50FN_ckoJi4MT2fjI"}}
