--- demo_sessions/1KkS6aAx9JHGTiHKmGN5iuKEsDZnIhHD0f8BCj64wtk.json
@@ -1,1 +1,0 @@
-{"session_id": "1KkS6aAx9JHGTiHKmGN5iuKEsDZnIhHD0f8BCj64wtk", "user_id": "local-user", "issued_at": "2025-08-13T03:45:48.840205+00:00", "expires_at": "2025-08-13T11:45:48.840205+00:00", "last_seen": "2025-08-13T03:45:48.851985+00:00", "ua_hash": "07d1d539047ef01990ffd8266d015775201aa150a054c0933cf358dd3c032775", "ip_net": null, "attrs": {"csrf": "ovidCiDXu2VGjHnL00wqO45DNYmCYcpsxFD2uSS2glw"}}
