--- demo_sessions/3_c3TVoXwcPNiyePUQBcjhR5_pfrNLQTi9ji2EdFnwU.json
@@ -1,1 +1,0 @@
-{"session_id": "3_c3TVoXwcPNiyePUQBcjhR5_pfrNLQTi9ji2EdFnwU", "user_id": "local-user", "issued_at": "2025-08-13T05:38:02.586145+00:00", "expires_at": "2025-08-13T13:38:02.586145+00:00", "last_seen": "2025-08-13T05:38:02.586145+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "Oj-DBsliyeyNbnkga4La1blmSUii2I-BSqw4UJF8li4"}}
