--- demo_sessions/9anqUfGgEVScpzLKi7DndGt0Q-v3cekE9-GcXrRVXQw.json
@@ -1,1 +1,0 @@
-{"session_id": "9anqUfGgEVScpzLKi7DndGt0Q-v3cekE9-GcXrRVXQw", "user_id": "local-user", "issued_at": "2025-08-13T03:47:53.095263+00:00", "expires_at": "2025-08-13T11:47:53.095263+00:00", "last_seen": "2025-08-13T03:47:53.142693+00:00", "ua_hash": "07d1d539047ef01990ffd8266d015775201aa150a054c0933cf358dd3c032775", "ip_net": null, "attrs": {"csrf": "PtIysshpkKyalnEM8CACEmKVBGgkZBPcZ0-JwxAA_GE"}}
