--- demo_sessions/AFOlSmX27-03W-7VtuBL_7XYiuNYqvdpTFU_0QB3GMc.json
@@ -1,1 +1,0 @@
-{"session_id": "AFOlSmX27-03W-7VtuBL_7XYiuNYqvdpTFU_0QB3GMc", "user_id": "local-user", "issued_at": "2025-08-13T05:44:56.610216+00:00", "expires_at": "2025-08-13T13:44:56.610216+00:00", "last_seen": "2025-08-13T05:44:56.610216+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "RGLbg2ZArNfwBHqZtBV0PYuDQTJ4BVG8FG2FWnSMVJE"}}
