--- demo_sessions/AzHS9O6ieqzPYmBIiqAKAQ54KHDVPPZFfBnanr-C5s4.json
@@ -1,1 +1,0 @@
-{"session_id": "AzHS9O6ieqzPYmBIiqAKAQ54KHDVPPZFfBnanr-C5s4", "user_id": "local-user", "issued_at": "2025-08-13T05:44:52.007577+00:00", "expires_at": "2025-08-13T13:44:52.007577+00:00", "last_seen": "2025-08-13T05:44:52.007577+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "7YMoFoKGmN7WdrVdTdnlVq7eu8d72SZAf6a0CWXcmPg"}}
