--- demo_sessions/BBrKESYUaAHPFR0HFnKYSXBh8qf8onTJnA4UdK3GWWc.json
@@ -1,1 +1,0 @@
-{"session_id": "BBrKESYUaAHPFR0HFnKYSXBh8qf8onTJnA4UdK3GWWc", "user_id": "local-user", "issued_at": "2025-08-13T05:24:32.799365+00:00", "expires_at": "2025-08-13T13:24:32.799365+00:00", "last_seen": "2025-08-13T05:32:39.738577+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "bTdREhhQI6YxPoOrJ_l8cb8d-CJzimBdKuKDqIzRBp4"}}
