--- demo_sessions/Ce5nB64oAlYkVM_IyyFsWS9nXzxejHfpId4pUTXNCuI.json
@@ -1,1 +1,0 @@
-{"session_id": "Ce5nB64oAlYkVM_IyyFsWS9nXzxejHfpId4pUTXNCuI", "user_id": "local-user", "issued_at": "2025-08-13T05:38:49.145333+00:00", "expires_at": "2025-08-13T13:38:49.145333+00:00", "last_seen": "2025-08-13T05:38:49.165302+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "aPEot4bPdLmv960DIvUa8FtcaK7uHcIZ9hqUQmUk4dE"}}
