--- demo_sessions/FN28V9RU9fnaW3REvWkT7XswCH0kdNuLXa2d7JhEuMc.json
@@ -1,1 +1,0 @@
-{"session_id": "FN28V9RU9fnaW3REvWkT7XswCH0kdNuLXa2d7JhEuMc", "user_id": "local-user", "issued_at": "2025-08-13T05:44:57.898592+00:00", "expires_at": "2025-08-13T13:44:57.898592+00:00", "last_seen": "2025-08-13T05:44:57.898592+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "I6nQ88mMThNJ_KFhC5LX1ZdHoiQpm881QsayAq5AoLc"}}
