--- demo_sessions/G2h1Q0YQISLuivwWZGDano-Rjb4olv_LrghKXPWZiIE.json
@@ -1,1 +1,0 @@
-{"session_id": "G2h1Q0YQISLuivwWZGDano-Rjb4olv_LrghKXPWZiIE", "user_id": "local-user", "issued_at": "2025-08-13T05:07:15.466208+00:00", "expires_at": "2025-08-13T13:07:15.466208+00:00", "last_seen": "2025-08-13T05:07:16.139019+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "6HFMyGUXxXaMDIA73Wik0DfX-Z6D2NzkCmOWyG3IZVc"}}
