--- demo_sessions/HUICCQ1UMTIhYGR3DN7n-64-vGop4ex-zwkacBx6i4M.json
@@ -1,1 +1,0 @@
-{"session_id": "HUICCQ1UMTIhYGR3DN7n-64-vGop4ex-zwkacBx6i4M", "user_id": "local-user", "issued_at": "2025-08-13T03:44:40.412255+00:00", "expires_at": "2025-08-13T11:44:40.412255+00:00", "last_seen": "2025-08-13T03:44:40.423830+00:00", "ua_hash": "07d1d539047ef01990ffd8266d015775201aa150a054c0933cf358dd3c032775", "ip_net": null, "attrs": {"csrf": "F5uLZAtX_6uRSEPW4HKcFIUxJdnupewk1tq7iJEBPhc"}}
