--- demo_sessions/KopvkE2_0GYVjWssoO-_hKdm6n4ziBcQ-VsVz-i8afo.json
@@ -1,1 +1,0 @@
-{"session_id": "KopvkE2_0GYVjWssoO-_hKdm6n4ziBcQ-VsVz-i8afo", "user_id": "local-user", "issued_at": "2025-08-13T05:07:06.633485+00:00", "expires_at": "2025-08-13T13:07:06.633485+00:00", "last_seen": "2025-08-13T05:07:06.655275+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "V7oqPya_Y8suvCxyewqqmZP-cRsdVb8myRsQF289wMc"}}
