--- demo_sessions/L77muksQaLURZ-0eO5ueb98-reZ0YEQ22114UMl_sxk.json
@@ -1,1 +1,0 @@
-{"session_id": "L77muksQaLURZ-0eO5ueb98-reZ0YEQ22114UMl_sxk", "user_id": "local-user", "issued_at": "2025-08-13T05:44:56.610381+00:00", "expires_at": "2025-08-13T13:44:56.610381+00:00", "last_seen": "2025-08-13T05:44:56.610381+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "kWEe5O3-Pdb3TI0NiNhyZzJrw0vacedL_PuBM3M4q1A"}}
