--- demo_sessions/NdkBV3eO3Ds7ly0FihGI_5cCVr_SbTX94K_h3TQcExE.json
@@ -1,1 +1,0 @@
-{"session_id": "NdkBV3eO3Ds7ly0FihGI_5cCVr_SbTX94K_h3TQcExE", "user_id": "local-user", "issued_at": "2025-08-13T05:38:49.145126+00:00", "expires_at": "2025-08-13T13:38:49.145126+00:00", "last_seen": "2025-08-13T05:38:49.145126+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "7zq0lTg5b6zXfh6Z-ti8UlDVBLgH3rE3xB-di4GSrsg"}}
