--- demo_sessions/PfjNC7Y5TgphYdQkQfXlMSERK_MfwC3alICN1xE1sVY.json
@@ -1,1 +1,0 @@
-{"session_id": "PfjNC7Y5TgphYdQkQfXlMSERK_MfwC3alICN1xE1sVY", "user_id": "local-user", "issued_at": "2025-08-13T05:23:32.958543+00:00", "expires_at": "2025-08-13T13:23:32.958543+00:00", "last_seen": "2025-08-13T05:24:32.800648+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "uuTkJRqU1GbZDodY9Y01eROUCFruMtyvV6WUORU7hgA"}}
