--- demo_sessions/U8EJjtMYK9O6cwa8fNZ1wGH68mT2WXDr8B7RX350Kps.json
@@ -1,1 +1,0 @@
-{"session_id": "U8EJjtMYK9O6cwa8fNZ1wGH68mT2WXDr8B7RX350Kps", "user_id": "local-user", "issued_at": "2025-08-13T05:23:13.933988+00:00", "expires_at": "2025-08-13T13:23:13.933988+00:00", "last_seen": "2025-08-13T05:23:32.959737+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "w-fGFlqIinWeZtTSBE4JrNvMSOn_FQCtXbLYe9AO-HI"}}
