--- demo_sessions/X6R09R_s50G5L_NFocIB3efWacQS7sFP3QYvUjmAfiQ.json
@@ -1,1 +1,0 @@
-{"session_id": "X6R09R_s50G5L_NFocIB3efWacQS7sFP3QYvUjmAfiQ", "user_id": "local-user", "issued_at": "2025-08-13T05:44:57.898854+00:00", "expires_at": "2025-08-13T13:44:57.898854+00:00", "last_seen": "2025-08-13T05:44:57.898854+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "u7q0_x6opQsew5-xdBD8_PunoTCNB9aLy86zT_G2LAw"}}
