--- demo_sessions/bFWtmUj-xUGxokz5CrSE4FQ9nQo19rrdzS2a52HyJnU.json
@@ -1,1 +1,0 @@
-{"session_id": "bFWtmUj-xUGxokz5CrSE4FQ9nQo19rrdzS2a52HyJnU", "user_id": "local-user", "issued_at": "2025-08-13T05:23:13.340735+00:00", "expires_at": "2025-08-13T13:23:13.340735+00:00", "last_seen": "2025-08-13T05:23:13.934828+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "otJ3OhQWQiPsB5eICmme2FO3eSvNWCsH0bSDrfZbWx8"}}
