--- demo_sessions/gmD2he0eM5Y6wgPXrfmOySJyWSQtMrr_bsUAH1ivJ9g.json
@@ -1,1 +1,0 @@
-{"session_id": "gmD2he0eM5Y6wgPXrfmOySJyWSQtMrr_bsUAH1ivJ9g", "user_id": "local-user", "issued_at": "2025-08-13T03:46:29.554855+00:00", "expires_at": "2025-08-13T11:46:29.554855+00:00", "last_seen": "2025-08-13T03:46:29.565830+00:00", "ua_hash": "07d1d539047ef01990ffd8266d015775201aa150a054c0933cf358dd3c032775", "ip_net": null, "attrs": {"csrf": "-8bV0u-1yFOSIhXfenb2RCZzTG89_1RDlJgZjKKdzXc"}}
