--- demo_sessions/l4WRc81mHYKlg7CLQYfzI6i0HYfyRU3jjV5CID-iwCc.json
@@ -1,1 +1,0 @@
-{"session_id": "l4WRc81mHYKlg7CLQYfzI6i0HYfyRU3jjV5CID-iwCc", "user_id": "local-user", "issued_at": "2025-08-13T05:07:07.374049+00:00", "expires_at": "2025-08-13T13:07:07.374049+00:00", "last_seen": "2025-08-13T05:07:15.439004+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "n6somAI9A1Vv364xLoEWmL2dRJPYd4S7_1ufbNZFt2k"}}
