--- demo_sessions/oEIdjffU7Sx2TpsdI1BiiM1hixBJW7i7lJ7PCjRZ2sM.json
@@ -1,1 +1,0 @@
-{"session_id": "oEIdjffU7Sx2TpsdI1BiiM1hixBJW7i7lJ7PCjRZ2sM", "user_id": "local-user", "issued_at": "2025-08-13T05:32:39.770490+00:00", "expires_at": "2025-08-13T13:32:39.770490+00:00", "last_seen": "2025-08-13T05:32:41.716100+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "QiHJENoo2piHXYpzm0rsVWWsAZlLVS7JYk9OHXY5iVM"}}
