--- demo_sessions/pj1Q6XZnQI09fr-fwWsD4XjpU1L4DUZBX2Kpu3bHU8g.json
@@ -1,1 +1,0 @@
-{"session_id": "pj1Q6XZnQI09fr-fwWsD4XjpU1L4DUZBX2Kpu3bHU8g", "user_id": "local-user", "issued_at": "2025-08-13T05:32:42.056605+00:00", "expires_at": "2025-08-13T13:32:42.056605+00:00", "last_seen": "2025-08-13T05:32:42.056605+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "8IlyT3gN12TeSXFzljGu7h8k1f1Qpx-loUByjB_O0uc"}}
