--- demo_sessions/vAjyC6-kU7IUig1UE476KZyH8Je1lHeSDyZx-tBIGe8.json
@@ -1,1 +1,0 @@
-{"session_id": "vAjyC6-kU7IUig1UE476KZyH8Je1lHeSDyZx-tBIGe8", "user_id": "local-user", "issued_at": "2025-08-13T05:38:02.585785+00:00", "expires_at": "2025-08-13T13:38:02.585785+00:00", "last_seen": "2025-08-13T05:38:48.614017+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "p1WGne26-_owYwaJwwVJlgPkpl3ABhh6QDKpjLTWpik"}}
