--- demo_sessions/wPkyoM-5r1YmkPcTjY7UJZjG3baxHGOf6CIuCexbhJE.json
@@ -1,1 +1,0 @@
-{"session_id": "wPkyoM-5r1YmkPcTjY7UJZjG3baxHGOf6CIuCexbhJE", "user_id": "local-user", "issued_at": "2025-08-13T03:44:54.000900+00:00", "expires_at": "2025-08-13T11:44:54.000900+00:00", "last_seen": "2025-08-13T03:44:54.011243+00:00", "ua_hash": "07d1d539047ef01990ffd8266d015775201aa150a054c0933cf358dd3c032775", "ip_net": null, "attrs": {"csrf": "2mjGxpSrUJzjouiQafdXpwG1Rl4wjzWxGL7WYNP7FLk"}}
