--- demo_sessions/xPnsVf-gPvxT-My7NoIceLwJmqZiaM7NxZWXdh_SRTw.json
@@ -1,1 +1,0 @@
-{"session_id": "xPnsVf-gPvxT-My7NoIceLwJmqZiaM7NxZWXdh_SRTw", "user_id": "local-user", "issued_at": "2025-08-13T05:32:42.056855+00:00", "expires_at": "2025-08-13T13:32:42.056855+00:00", "last_seen": "2025-08-13T05:38:01.862537+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "bVtfvsIRTtCEpqWZEIyIJkv2h2AFbOLxj1sNxF9NIxo"}}
