--- demo_sessions/zgROWi3tC1rxHBju85jRvimKq6HY2RPXxPhNeXDsexU.json
@@ -1,1 +1,0 @@
-{"session_id": "zgROWi3tC1rxHBju85jRvimKq6HY2RPXxPhNeXDsexU", "user_id": "local-user", "issued_at": "2025-08-13T05:44:52.010694+00:00", "expires_at": "2025-08-13T13:44:52.010694+00:00", "last_seen": "2025-08-13T05:44:52.010694+00:00", "ua_hash": "9559d1e35f81d5056027cc076b5180ea0a8606127d56fdf2f861c9185cc17cb7", "ip_net": null, "attrs": {"csrf": "5jb7pnKFpq-2FEDUwt1-rJ7x57ahpuvSUYtkIeu3A8Y"}}
