+++ src/knowledge_web/templates/pages/index.html
@@ -1,0 +1,45 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="utf-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1" />
+  <title>Deployable UI Demo</title>
+  <link rel="stylesheet" href="/static/ui/css/theme.css" />
+  <link rel="stylesheet" href="/static/ui/css/style.css" />
+</head>
+<body>
+  <div class="app">
+    <div class="app-header">
+      <div class="left">
+        <div class="menu">
+          <button id="menu-trigger" class="menu-trigger" aria-haspopup="true" aria-expanded="false">Menu ▾</button>
+          <div id="menu-dropdown" class="menu-dropdown" role="menu" aria-hidden="true">
+            <button class="menu-item" data-action="spawn-left" role="menuitem">New Left Window</button>
+            <button class="menu-item" data-action="spawn-right" role="menuitem">New Right Window</button>
+            <div class="menu-sep"></div>
+            <button class="menu-item" data-action="spawn-dockable-modal" role="menuitem">Dockable Modal</button>
+            <button class="menu-item" data-action="spawn-simple-modal" role="menuitem">Simple Modal</button>
+            <div class="menu-sep"></div>
+            <button class="menu-item" data-action="spawn-showcase" role="menuitem">UI Showcase</button>
+            <div class="menu-sep"></div>
+            <button class="menu-item" data-action="spawn-editor" role="menuitem">Text Editor</button>
+            <button class="menu-item" data-action="spawn-listview" role="menuitem">List View Demo</button>
+            <button class="menu-item" data-action="spawn-chat" role="menuitem">Chat</button>
+            <button class="menu-item" data-action="spawn-upload" role="menuitem">File Upload Demo</button>
+          </div>
+        </div>
+      </div>
+      <div class="brand"><strong>Deployable UI Demo</strong></div>
+      <div class="right"></div>
+    </div>
+
+    <div class="columns" id="columns">
+      <div class="col" id="col-left"></div>
+      <div class="splitter" id="splitter" role="separator" aria-orientation="vertical" aria-label="Resize Columns"></div>
+      <div class="col" id="col-right"></div>
+    </div>
+  </div>
+
+  <script type="module" src="/static/js/main.js"></script>
+</body>
+</html>
