--- src/templates/pages/index.html
@@ -1,53 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="utf-8" />
-  <meta
-    name="viewport"
-    content="width=device-width,initial-scale=1,viewport-fit=cover"
-  />
-  <title>Deployable Knowledge — Demo</title>
-
-  <!-- Optional: your theme/base styles -->
-<link rel="stylesheet" href="/ui/css/ui.css" />
-<link rel="stylesheet" href="/ui/css/style.css" />
-<link rel="stylesheet" href="/ui/css/theme.css" />
-
-  <!-- Minimal inline styles so the windows look decent even without your CSS -->
-  <style>
-    :root {
-      --bg: #0b0f14;
-      --fg: #dbe2ea;
-      --panel: #151b22;
-      --panel-border: #243140;
-      --accent: #5ec3ff;
-    }
-    html, body { height:100%; margin:0; background:var(--bg); color:var(--fg); font:14px/1.4 system-ui, -apple-system, Segoe UI, Roboto, sans-serif; }
-    #app { position:relative; width:100%; height:100%; overflow:hidden; }
-    /* lightweight “window” look; your UI lib likely has its own */
-    .window { position:absolute; background:var(--panel); border:1px solid var(--panel-border); border-radius:10px; box-shadow:0 6px 24px rgba(0,0,0,.35); overflow:hidden; }
-    .window .titlebar { padding:8px 12px; border-bottom:1px solid var(--panel-border); font-weight:600; }
-    .window-body { height:100%; }
-    /* chat bubbles used by our minimal chat window renderer */
-    .bubble { margin:8px 0; padding:8px 10px; border-radius:10px; background:#0f141a; border:1px solid var(--panel-border); }
-    .bubble.user { background:#0e1a24; border-color:#22425a; }
-    .bubble.assistant { background:#131a11; border-color:#224a2a; }
-    .bubble .role { font-size:12px; opacity:.7; margin-bottom:4px; }
-    .btn { background:var(--accent); color:#001018; border:none; padding:6px 10px; border-radius:8px; cursor:pointer; }
-    textarea.chat-input { flex:1; min-height:40px; max-height:120px; width:100%; }
-  </style>
-</head>
-<body>
-  <div id="app"></div>
-
-  <!-- Boot the app (ES modules). This file imports from /src/ui/js/* and /src/sdk/sdk.js -->
-  <script type="module" src="/src/app/boot.js"></script>
-
-  <!-- Optional: safety note for no-JS -->
-  <noscript>
-    <div style="padding:1rem; color:#000; background:#ffc;">
-      This app needs JavaScript enabled.
-    </div>
-  </noscript>
-</body>
-</html>
