--- src/ui-compat/compat_components.js
@@ -1,2 +1,0 @@
-// compat_components.js — re-exports components + bus from the new UI library
-export { createItemList, registerComponent, getComponent, bus } from "/ui/js/components.js";
