--- src/ui-compat/compat_ui.js
@@ -1,3 +1,0 @@
-// compat_ui.js — re-exports from the new UI library
-// Keeps old import paths stable for legacy controllers.
-export { el, Field, fieldRow } from "/ui/js/ui.js";
