--- src/ui-compat/windows/window_documents.js
@@ -1,2 +1,0 @@
-// window_documents.js — wrapper (kept for legacy imports)
-export { } from "../windows.js";
