--- src/ui-compat/windows/window_persona.js
@@ -1,2 +1,0 @@
-// window_persona.js — wrapper (legacy)
-export { } from "../windows.js";
