--- src/ui-compat/windows/window_prompt_editor.js
@@ -1,2 +1,0 @@
-// window_prompt_editor.js — wrapper (legacy)
-export { } from "../windows.js";
