--- src/ui-compat/windows/window_search.js
@@ -1,2 +1,0 @@
-// window_search.js — wrapper (legacy)
-export { } from "../windows.js";
