--- src/ui-compat/windows/window_segment_view.js
@@ -1,2 +1,0 @@
-// window_segment_view.js — wrapper (legacy)
-export { } from "../windows.js";
