--- src/ui-compat/windows/window_segments.js
@@ -1,2 +1,0 @@
-// window_segments.js — wrapper (legacy)
-export { } from "../windows.js";
