--- src/ui-compat/windows/window_sessions.js
@@ -1,2 +1,0 @@
-// window_sessions.js — wrapper (legacy)
-export { } from "../windows.js";
