--- src/ui/css/style.css
@@ -1,1 +1,0 @@
-/* placeholder for additional styles */
