--- src/ui/css/theme.css
@@ -1,1 +1,0 @@
-/* placeholder theme */
